async get_reply in chain tracker, split into its own module
